Refactored table component
